+++ src/Serializer/MapCell.php
@@ -24,7 +24,8 @@ final class MapCell
     public function __construct(
         public readonly string|int|null $column = null,
         public readonly ?string $cast = null,
-        public readonly array $options = []
+        public readonly array $options = [],
+        public readonly bool $ignore = false,
     ) {
     }
 }